+++ src/components/globals/index.ts
@@ -1,3 +1,3 @@
-export {default as Header} from './Header';
-export {default as Footer} from './Footer';
+export {default as Header} from '../molecules/Header/Header';
+export {default as Footer} from '../molecules/Footer/Footer';
 export {default as Layout} from './Layout';
