+++ src/components/molecules/Header/Header.module.scss
@@ -10,8 +10,8 @@
   background-color: rgba(255, 255, 255, 1);
   /* background-color: rgba(255, 255, 255, 0.7); */
   -webkit-box-shadow: 0 10px 6px -6px #777;
-       -moz-box-shadow: 0 10px 6px -6px #777;
-            box-shadow: 0 10px 6px -6px #777;
+  -moz-box-shadow: 0 10px 6px -6px #777;
+  box-shadow: 0 10px 6px -6px #777;
   z-index: 10;
 }
 .site_title{
