+++ src/components/molecules/Header/Header.tsx
@@ -1,0 +1,41 @@
+import { CSSProperties, FC, useEffect, useState } from 'react';
+import Link from 'next/link';
+import styles from './Header.module.scss';
+import ClassNames from 'classnames';
+import { ThemeContext, useThemeContext } from 'context/context';
+
+const scrollTop = (): number => Math.max(window.pageYOffset, document.documentElement.scrollTop, document.body.scrollTop);
+
+const Header: FC = () => {
+  const [isTop, setIsTop] = useState<boolean>(true);
+  const themeNames: ThemeContext = useThemeContext();
+  const HeaderStyle: string = ClassNames(styles.header, {
+    [styles.dark]: themeNames.themeName === 'dark'
+  })
+  
+  const onScroll = (): void => scrollTop() >= 5 ? setIsTop(false) : setIsTop(true);
+  
+  useEffect(() => {
+    document.addEventListener("scroll", onScroll);
+    return (): void => document.removeEventListener("scroll", onScroll);
+  });
+  
+  const scrollStyle: CSSProperties = isTop ? {} : {width: '30%'};
+  
+  return (
+    <header className={HeaderStyle}>
+      <h1 className={styles.site_title}>
+        <Link href="/">
+          <a id="header-logo" style={scrollStyle} href="/">
+            {themeNames.themeName === 'dark' ?
+              <img src={"/azerbaijapan-logo-dark.png"} alt=""/> :
+              <img src={"/azerbaijapan.png"} alt=""/>
+            }
+          </a>
+        </Link>
+      </h1>
+    </header>
+  )
+}
+
+export default Header
