+++ src/pages/posts/ja/[id].tsx
@@ -1,7 +1,9 @@
-import { wpBaseUrl, wpGenerateNextAndPrevArray, wpGetPostDataById } from 'lib/post';
-import { GetServerSideProps, GetServerSidePropsContext } from "next";
+import { wpBaseUrl, wpGenerateNextAndPrevArray } from 'lib/post';
+import { GetServerSideProps, GetServerSidePropsContext, NextPage } from "next";
 import PostPage from "components/template/PostPage/PostPage";
 import { fetchWithCache } from "lib/helpers";
+import { wpGetCatNamesById } from "lib/category";
+import { wpGetTagNamesById } from "lib/tags";
 
 // postの中のcssはglobal.cssに記載
 
@@ -21,7 +23,7 @@ export const getServerSideProps: GetServerSideProps = async (context: GetServerS
   }
   await Promise.all([
     (async () => {
-      postData = await wpGetPostDataById('ja', Number(context.query.id));
+      postData = await fetchWithCache(`${wpBaseUrl}/ja/wp-json/wp/v2/posts/${Number(context.query.id)}?_fields=id,acf,title,date,modified,content,meta,categories,category_name,tags,translate_group`);
     })(),
     (async () => {
       nextAndPrev = await wpGenerateNextAndPrevArray('ja', Number(context.query.id));
@@ -51,6 +53,13 @@ export const getServerSideProps: GetServerSideProps = async (context: GetServerS
       tags['ru'] = await fetchWithCache(`${wpBaseUrl}/ja/wp-json/wp/v2/tags`)
     })(),
   ]);
+  
+  postData['cat_obj'] = await Promise.all(postData.categories.map(each => (
+    wpGetCatNamesById(each)
+  )))
+  postData['tags_obj'] = await Promise.all(postData.tags.map(each => (
+    wpGetTagNamesById(each)
+  )))
   return {
     props: {
       postData,
@@ -61,7 +70,7 @@ export const getServerSideProps: GetServerSideProps = async (context: GetServerS
   }
 }
 
-const Post = ({postData, categories, nextAndPrev, tags}) => {
+const Post: NextPage = ({postData, categories, nextAndPrev, tags}) => {
   
   return (
     <PostPage postData={postData} categories={categories} nextAndPrev={nextAndPrev} tags={tags}/>
